Added BCD_SECT specifiers in Hugh's new BCD code. Added a few int4 casts for PalmOS 68k compatibility; more of these will probably be needed -- must revisit all of bcd*. Changed termination condition in solver'a do_ridders code: the problem was that the value of (a+b)/ 2 is not necessarily within [a,b] in decimal, for example if a and b both start with 6, and a ends with 16, and b ends with 18, a+b will start with 12 and end with 3 (16+18=34; the final digit is rounded off), and (a+b)/2 will end with 15. (In binary, this cannot happen.) The new termination condition prevents the solver from looping endlessly when this occurs, but the right thing to do would be to calculate (a+b)/2 in a way that is robust against the round-off problem.

+++ common/bcdfloat2.cc
@@ -65,11 +65,11 @@ int BCDFloat2::_round25()
     if (d_[0] < 10)
         v = d_[P2-1] + (d_[P2] >= 5000);
     else if (d_[0] < 100)
-        v = (((d_[P2-1]+5)*3277)>>15)*10;
+        v = (((((int4) d_[P2-1])+5)*3277)>>15)*10;
     else if (d_[0] < 1000)
-        v = (((d_[P2-1]+50)*5243)>>19)*100;
+        v = (((((int4) d_[P2-1])+50)*5243)>>19)*100;
     else
-        v = (((d_[P2-1]+500)*8389)>>23)*1000;
+        v = (((((int4) d_[P2-1])+500)*8389)>>23)*1000;
 
     i = P2-1;
     while (v >= BASE)
@@ -824,7 +824,7 @@ bool BCDFloat2::sqrt(const BCDFloat2* a, BCDFloat2* r)
     
     BCDFloat2 rr(r1);
     BCDFloat2 t, t2;
-    BCDFloat2 two(2U);
+    BCDFloat2 two(2);
     if (res)
     {
         if (!rr.isSpecial())
